+++ examples/data/id/seq.rs
@@ -1,4 +1,4 @@
-// devela/examples/data/id_seq.rs
+// devela/examples/data/id/seq.rs
 //
 //! Shows a sequential ID generator made with the [`id_seq!`] macro.
 //!