+++ src/code/util/_test/mod.rs
@@ -1,4 +1,3 @@
 // devela/src/code/util/_test/mod.rs
 
-// mod cdbg;
 mod symbols;