+++ src/data/id/seq.rs
@@ -5,7 +5,7 @@
 
 #[doc = crate::_tags!(construction uid)]
 /// A macro for constructing a unique sequential identifier generator.
-#[doc = crate::_doc_meta!{location("data/uid")}]
+#[doc = crate::_doc_meta!{location("data/id")}]
 ///
 /// It generates the necessary static instances dynamically.
 ///
@@ -26,20 +26,19 @@
 /// }
 /// assert_eq![AppId1::new_fast(), None];
 /// ```
-///
-/// See also [`IdSeqUsizeExample`][crate::IdSeqUsizeExample],
-/// generated in the [id_seq][crate::_doc::examples::id_seq] example.
+/// See also [`IdSeqU64Example`][crate::IdSeqU64Example].
 #[macro_export]
 #[cfg_attr(cargo_primary_package, doc(hidden))]
 macro_rules! id_seq {
     (
-        // $name: the name of the sequential ID generator. E.g. AppId1.
-        // $prim: the underlying primitive type. E.g. u64.
-        $name:ident,
+        $(#[$attr:meta])*
+        $vis:vis $name:ident,
         $prim:ty
     ) => {
         $crate::paste! {
-            $crate::id_seq![
+            $crate::id_seq![%define
+                $(#[$attr])*,  // the attributes
+                $vis, // the visibility
                 $name, // the name of the ID generator.
                 $name, // the name of the ID generator (as a type).
                 [<$name:upper>], // the name of the static.
@@ -48,25 +47,16 @@ macro_rules! id_seq {
             ];
         }
     };
-
-    ($name:ident, $tname:ty, $static:ident, $prim:ty, $atomic:ident) => {
+    (%define $(#[$attr:meta])*, $vis:vis, $name:ident, $tname:ty,
+     $static:ident, $prim:ty, $atomic:ident) => {
         /// A static atomic counter used to generate unique sequential identifiers of type `$prim`.
         static $static: ::core::sync::atomic::$atomic =
             ::core::sync::atomic::$atomic::new(<$prim>::MIN + 1);
 
-        #[doc = concat!("A unique sequential identifier `", stringify!($prim), "` generator.")]
-        ///
-        #[doc = concat!("The counter starts at [`", stringify!($prim), "::MIN`]` + 1`,")]
-        /// and increments with each new identifier generated.
-        ///
-        /// The implementation guards against wrap-around
-        #[doc = concat!("after [`", stringify!($prim), "::MAX`],")]
-        /// by returning `None` in the checked methods, or panicking in the unchecked methods.
-        ///
-        /// See also the [`id_seq`] macro.
+        $(#[$attr])*
         #[derive(Debug)]
         #[must_use]
-        pub struct $name {
+        $vis struct $name {
             id: $prim,
         }
 
@@ -79,7 +69,7 @@ macro_rules! id_seq {
             ///
             /// Returns `None` on overflow.
             #[must_use]
-            pub fn new() -> Option<Self> {
+            $vis fn new() -> Option<Self> {
                 Self::new_balanced()
             }
             #[doc = concat!("Generates some unique `", stringify!($name), "` ID.")]
@@ -88,7 +78,7 @@ macro_rules! id_seq {
             ///
             /// # Panics
             /// Panics on overflow.
-            pub fn new_unchecked() -> Self {
+            $vis fn new_unchecked() -> Self {
                 Self::new_balanced_unchecked()
             }
 
@@ -103,7 +93,7 @@ macro_rules! id_seq {
             ///
             /// [`SeqCst`]: core::sync::atomic::Ordering::SeqCst
             #[must_use]
-            pub fn new_strong() -> Option<Self> {
+            $vis fn new_strong() -> Option<Self> {
                 Self::new_custom(::core::sync::atomic::Ordering::SeqCst)
             }
             #[doc = concat!(
@@ -117,7 +107,7 @@ macro_rules! id_seq {
             /// Panics on overflow.
             ///
             /// [`SeqCst`]: core::sync::atomic::Ordering::SeqCst
-            pub fn new_strong_unchecked() -> Self {
+            $vis fn new_strong_unchecked() -> Self {
                 Self::new_custom_unchecked(::core::sync::atomic::Ordering::SeqCst)
             }
 
@@ -132,7 +122,7 @@ macro_rules! id_seq {
             ///
             /// [`AcqRel`]: core::sync::atomic::Ordering::AcqRel
             #[must_use]
-            pub fn new_balanced() -> Option<Self> {
+            $vis fn new_balanced() -> Option<Self> {
                 Self::new_custom(::core::sync::atomic::Ordering::AcqRel)
              }
             #[doc = concat!(
@@ -146,7 +136,7 @@ macro_rules! id_seq {
             /// Panics on overflow.
             ///
             /// [`AcqRel`]: core::sync::atomic::Ordering::AcqRel
-            pub fn new_balanced_unchecked() -> Self {
+            $vis fn new_balanced_unchecked() -> Self {
                 Self::new_custom_unchecked(::core::sync::atomic::Ordering::AcqRel)
             }
 
@@ -161,9 +151,9 @@ macro_rules! id_seq {
             ///
             /// [`Relaxed`]: core::sync::atomic::Ordering::Relaxed
             #[must_use]
-            pub fn new_fast() -> Option<Self> {
+            $vis fn new_fast() -> Option<Self> {
                 Self::new_custom(::core::sync::atomic::Ordering::Relaxed)
-             }
+            }
             #[doc = concat!(
                 "Generates a unique `", stringify!($name), "` ID with [`Relaxed`] ordering.")
             ]
@@ -175,7 +165,7 @@ macro_rules! id_seq {
             /// Panics on overflow.
             ///
             /// [`Relaxed`]: core::sync::atomic::Ordering::Relaxed
-            pub fn new_fast_unchecked() -> Self {
+            $vis fn new_fast_unchecked() -> Self {
                 Self::new_custom_unchecked(::core::sync::atomic::Ordering::Relaxed)
             }
 
@@ -187,7 +177,7 @@ macro_rules! id_seq {
             /// even at the cost of performance.
             ///
             /// [`SeqCst`]: core::sync::atomic::Ordering::SeqCst
-            pub fn iter_strong() -> impl Iterator<Item = $name> {
+            $vis fn iter_strong() -> impl Iterator<Item = $name> {
                 ::core::iter::from_fn(|| Self::new_strong())
             }
             /// Iterator over generated IDs with `SeqCst` ordering.
@@ -199,7 +189,7 @@ macro_rules! id_seq {
             /// Panics on overflow.
             ///
             /// [`SeqCst`]: core::sync::atomic::Ordering::SeqCst
-            pub fn iter_strong_unchecked() -> impl Iterator<Item = $name> {
+            $vis fn iter_strong_unchecked() -> impl Iterator<Item = $name> {
                 ::core::iter::from_fn(|| Some(Self::new_strong_unchecked()))
             }
 
@@ -209,7 +199,7 @@ macro_rules! id_seq {
             /// ensuring consistent visibility across threads.
             ///
             /// [`AcqRel`]: core::sync::atomic::Ordering::AcqRel
-            pub fn iter_balanced() -> impl Iterator<Item = $name> {
+            $vis fn iter_balanced() -> impl Iterator<Item = $name> {
                 ::core::iter::from_fn(|| Self::new_balanced())
             }
             /// Iterator over generated IDs with `AcqRel` ordering.
@@ -221,7 +211,7 @@ macro_rules! id_seq {
             /// Panics on overflow.
             ///
             /// [`AcqRel`]: core::sync::atomic::Ordering::AcqRel
-            pub fn iter_balanced_unchecked() -> impl Iterator<Item = $name> {
+            $vis fn iter_balanced_unchecked() -> impl Iterator<Item = $name> {
                 ::core::iter::from_fn(|| Some(Self::new_balanced_unchecked()))
             }
 
@@ -231,7 +221,7 @@ macro_rules! id_seq {
             /// where memory ordering is not a concern.
             ///
             /// [`Relaxed`]: core::sync::atomic::Ordering::Relaxed
-            pub fn iter_fast() -> impl Iterator<Item = $name> {
+            $vis fn iter_fast() -> impl Iterator<Item = $name> {
                 ::core::iter::from_fn(|| Self::new_fast())
             }
             /// Iterator over generated IDs with `Relaxed` ordering.
@@ -243,7 +233,7 @@ macro_rules! id_seq {
             /// Panics on overflow.
             ///
             /// [`Relaxed`]: core::sync::atomic::Ordering::Relaxed
-            pub fn iter_fast_unchecked() -> impl Iterator<Item = $name> {
+            $vis fn iter_fast_unchecked() -> impl Iterator<Item = $name> {
                 ::core::iter::from_fn(|| Some(Self::new_fast_unchecked()))
             }
 
@@ -254,7 +244,7 @@ macro_rules! id_seq {
             ///
             /// The value is guaranteed to be a valid sequential identifier, from
             #[doc = concat!("`", stringify!($prim) ,"::MIN` to `", stringify!($prim), "::MAX`.")]
-            pub fn value(&self) -> $prim {
+            $vis fn value(&self) -> $prim {
                 self.id
             }
 
@@ -262,7 +252,7 @@ macro_rules! id_seq {
             ///
             /// Alias of [`generated_ids_balanced`][Self::generated_ids_balanced].
             #[must_use]
-            pub fn generated_ids() -> $prim {
+            $vis fn generated_ids() -> $prim {
                 Self::generated_ids_balanced()
             }
             /// Returns the number of IDs generated so far with [`SeqCst`] ordering.
@@ -272,7 +262,7 @@ macro_rules! id_seq {
             ///
             /// [`SeqCst`]: core::sync::atomic::Ordering::SeqCst
             #[must_use]
-            pub fn generated_ids_strong() -> $prim {
+            $vis fn generated_ids_strong() -> $prim {
                 let current_id = $static.load(::core::sync::atomic::Ordering::SeqCst);
                 current_id - 1
             }
@@ -283,7 +273,7 @@ macro_rules! id_seq {
             ///
             /// [`Acquire`]: core::sync::atomic::Ordering::Acquire
             #[must_use]
-            pub fn generated_ids_balanced() -> $prim {
+            $vis fn generated_ids_balanced() -> $prim {
                 let current_id = $static.load(::core::sync::atomic::Ordering::Acquire);
                 current_id - 1
             }
@@ -294,7 +284,7 @@ macro_rules! id_seq {
             ///
             /// [`Relaxed`]: core::sync::atomic::Ordering::Relaxed
             #[must_use]
-            pub fn generated_ids_fast() -> $prim {
+            $vis fn generated_ids_fast() -> $prim {
                 let current_id = $static.load(::core::sync::atomic::Ordering::Relaxed);
                 current_id - 1
             }
@@ -303,7 +293,7 @@ macro_rules! id_seq {
             ///
             /// Alias of [`remaining_ids_balanced`][Self::remaining_ids_balanced].
             #[must_use]
-            pub fn remaining_ids() -> $prim {
+            $vis fn remaining_ids() -> $prim {
                 Self::remaining_ids_balanced()
             }
             /// Returns the number of remaining IDs with [`SeqCst`] ordering.
@@ -313,7 +303,7 @@ macro_rules! id_seq {
             ///
             /// [`SeqCst`]: core::sync::atomic::Ordering::SeqCst
             #[must_use]
-            pub fn remaining_ids_strong() -> $prim {
+            $vis fn remaining_ids_strong() -> $prim {
                 let current_id = $static.load(::core::sync::atomic::Ordering::SeqCst);
                 <$prim>::MAX - current_id
             }
@@ -324,7 +314,7 @@ macro_rules! id_seq {
             ///
             /// [`Acquire`]: core::sync::atomic::Ordering::Acquire
             #[must_use]
-            pub fn remaining_ids_balanced() -> $prim {
+            $vis fn remaining_ids_balanced() -> $prim {
                 let current_id = $static.load(::core::sync::atomic::Ordering::Acquire);
                 <$prim>::MAX - current_id
             }
@@ -335,7 +325,7 @@ macro_rules! id_seq {
             ///
             /// [`Relaxed`]: core::sync::atomic::Ordering::Relaxed
             #[must_use]
-            pub fn remaining_ids_fast() -> $prim {
+            $vis fn remaining_ids_fast() -> $prim {
                 let current_id = $static.load(::core::sync::atomic::Ordering::Relaxed);
                 <$prim>::MAX - current_id
             }